working on styling for T&C
