+++ mocks/data/cms.tsx
@@ -357,146 +357,6 @@ export const mockedConfigurationRulesGuidanceProps = {
   ),
 };
 
-export const mockedTermsAndConditions = {
-  acceptingTerms: {
-    content: (
-      <>
-        <Typography variant="h5" gutterBottom>
-          Accepting Terms and Conditions
-        </Typography>
-        <Typography>
-          By clicking on "Accept" below, you agree to the terms and conditions
-          outlined on this page.
-        </Typography>
-      </>
-    ),
-  },
-  changesToTerms: {
-    content: (
-      <>
-        <Typography variant="h5" gutterBottom>
-          Changes to Terms and Conditions
-        </Typography>
-        <Typography>
-          The terms and conditions for SOURSD may change from time to time.
-          Please review the latest version before using our product.
-        </Typography>
-      </>
-    ),
-  },
-  usingOurProduct: {
-    content: (
-      <>
-        <Typography variant="h5" gutterBottom>
-          Using Our Product
-        </Typography>
-        <Typography>
-          The SOURSD product is designed to help you manage your sensitive data
-          more effectively.
-        </Typography>
-      </>
-    ),
-  },
-  generalRestrictions: {
-    content: (
-      <>
-        <Typography variant="h5" gutterBottom>
-          General Restrictions
-        </Typography>
-        <Typography>
-          SOURSD is not a replacement for professional legal advice. You should
-          consult with a qualified legal professional before using the product.
-        </Typography>
-      </>
-    ),
-  },
-  contentPolicy: {
-    content: (
-      <>
-        <Typography variant="h5" gutterBottom>
-          Content Policy
-        </Typography>
-        <Typography>
-          SOURSD collects and stores personal data in accordance with the
-          General Data Protection Regulation (GDPR) and other relevant data
-          protection regulations. The data collected includes your name, email
-          address, research project details, and affiliated Users' details.
-        </Typography>
-      </>
-    ),
-  },
-  yourRights: {
-    content: (
-      <>
-        <Typography variant="h5" gutterBottom>
-          Your Rights
-        </Typography>
-        <Typography>
-          You have the right to access, correct, update, or delete your personal
-          data. You also have the right to object to processing your data. You
-          can exercise your rights under the GDPR by contacting the GDPR data
-          protection officer.
-        </Typography>
-      </>
-    ),
-  },
-  copyrightPolicy: {
-    content: (
-      <>
-        <Typography variant="h5" gutterBottom>
-          Copyright Policy
-        </Typography>
-        <Typography>
-          SOURSD is protected by copyright law and other intellectual property
-          rights. The SOURSD logo and any other trademarks or service marks used
-          on this website are owned by the SOURSD team.
-        </Typography>
-      </>
-    ),
-  },
-  relationshipGuidelines: {
-    content: (
-      <>
-        <Typography variant="h5" gutterBottom>
-          Relationship Guidelines
-        </Typography>
-        <Typography>
-          SOURSD encourages users to establish trust and relationships with each
-          other. By creating a SOURSD account, you are agreeing to the SOURSD
-          relationship guidelines.
-        </Typography>
-      </>
-    ),
-  },
-  liabilityPolicy: {
-    content: (
-      <>
-        <Typography variant="h5" gutterBottom>
-          Liability Policy
-        </Typography>
-        <Typography>
-          SOURSD is not liable for any loss, damage, or inconvenience caused by
-          your use of the SOURSD product.
-        </Typography>
-      </>
-    ),
-  },
-  generalLegalTerms: {
-    content: (
-      <>
-        <Typography variant="h5" gutterBottom>
-          General Legal Terms
-        </Typography>
-        <Typography>
-          These terms and conditions are governed by the laws of the United
-          Kingdom. If you have any questions or concerns about these terms and
-          conditions, please contact our support team at [support@soursd.com].
-        </Typography>
-      </>
-    ),
-  },
-};
-
 export const mockedAbout = {
   infoTitle: "About SOURSD",
   aboutTitle: "Empowering Safe Data Access for Research and Innovation",
