+++ mocks/data/terms/business/10_disclaimers.md
@@ -1,0 +1,15 @@
+# 10. USE OF SOURSD AND DISCLAIMERS
+
+## 10.1 SOURSD is provided to you free of charge. We do not guarantee that SOURSD, or any content on it, will always be available, will be uninterrupted or error-free, or that defects will be corrected. Access to SOURSD is permitted on a temporary basis. We may suspend, withdraw, discontinue or change all or any part of SOURSD without notice. We will not be liable to you if for any reason SOURSD is unavailable at any time or for any period. We may update SOURSD and/or change the content on it at any time.
+
+## 10.2 SOURSD may contain inaccuracies or typographical errors. To the fullest extent permitted by applicable law, we make no representations about the reliability, availability, timeliness or accuracy of the content included on SOURSD and are not liable for and take no responsibility for any content or data uploaded by Users to SOURSD.
+
+## 10.3 To the fullest extent permitted by applicable law, SOURSD is provided “as is” and on an “as available” basis without warranties of any kind, expressed or implied. We disclaim all warranties, expressed or implied, including, but not limited to, implied warranties relating to satisfactory quality, fitness for a particular purpose, non-infringement, security or accurateness.
+
+## 10.4 We do not warrant or make any representation regarding the use or the results of the use of SOURSD in terms of their correctness, accuracy, reliability, or otherwise. To the fullest extent permitted by applicable law, we are not responsible for any decisions taken by a Data Custodian to allow Researchers and/or Research Organisations access to any TRE or SDE (or any data therein) and shall not be liable in the event that any Input Data is inaccurate. We do not make any recommendations to or decisions for Data Custodians in relation to access to any TRE or SDE.
+
+## 10.5 We will not be responsible for any loss or damage resulting from any machine, system, communications or power failure, technical breakdown, industrial dispute or other events or circumstances beyond our reasonable control leading to the total or partial unavailability of SOURSD.
+
+## 10.6 You are responsible for making all arrangements necessary for you to have access to SOURSD.
+
+## 10.7 We do not guarantee that SOURSD will be totally secure or free from bugs or viruses. You are responsible for configuring your information technology and computer programmes in order to access SOURSD and we recommend that you use your own virus protection software. We also assume no responsibility, and shall not be liable for, any damages to, or viruses that may infect, your computer equipment or other property as a result of your access to or use of SOURSD.
