+++ mocks/data/terms/business/11_suspension.md
@@ -1,0 +1,29 @@
+# 11. SUSPENSION AND TERMINATION
+
+## 11.1 Your use of SOURSD may, wholly or partly, be suspended if:
+
+### 11.1.1 we have concerns about the security of SOURSD;
+
+### 11.1.2 we suspect your use of SOURSD has been fraudulent or is being conducted in an unauthorised way; or
+
+### 11.1.3 we are required by law or regulation to do so.
+
+We will not be liable to you for the unavailability of SOURSD during such suspension.
+
+## 11.2 If you breach any of these Terms, we may immediately do any or all of the following (without limitation and without prejudice to any other right or remedy):
+
+### 11.2.1 issue a warning to you;
+
+### 11.2.2 temporarily or permanently remove any Input Data uploaded by you to SOURSD;
+
+### 11.2.3 temporarily or permanently withdraw your right to use SOURSD;
+
+### 11.2.4 suspend or terminate your Account;
+
+### 11.2.5 take any necessary legal action against you; and/or
+
+### 11.2.6 disclose such information to law enforcement authorities as we reasonably feel is necessary to do so.
+
+## 11.3 Without prejudice to any other rights of termination or suspension, we may terminate the operation of SOURSD at any time on written notice.
+
+## 11.4 Termination of your access to SOURSD is without prejudice to any other rights and remedies that we may have.
