+++ mocks/data/terms/business/12_changes.md
@@ -1,0 +1,3 @@
+# 12. CHANGES TO THESE TERMS
+
+## 12.1 We may make changes to these Terms from time to time (if, for example, there is a change in the law that means we need to change these Terms). Please check these Terms regularly to ensure that you understand the Terms that apply at the time that you access and use SOURSD. If you do not wish to continue using SOURSD following the changes to the Terms, you can cancel your agreement to these Terms by cancelling your Account and/or ceasing to use SOURSD.
