+++ mocks/data/terms/business/13_third_party.md
@@ -1,0 +1,3 @@
+# 13. THIRD PARTY CONTENT
+
+## 13.1 SOURSD may contain links to other websites which are not maintained by HDR UK. We assume no responsibility for the content of websites linked to and from SOURSD or the services provided through such websites (including identity verification services). We cannot guarantee that the links set out on our website will be accurate at the time of your access. Moreover, the sites pointed at by links have been independently developed and possibly maintained by person(s) over whom we have no control. We will not be liable for any loss or damage that may arise from your use of them.
