+++ mocks/data/terms/business/14_other.md
@@ -1,0 +1,15 @@
+# 14. OTHER IMPORTANT INFORMATION
+
+## 14.1 Each of the clauses of these Terms operates separately. If any court or relevant authority decides that any of them are unlawful or unenforceable, the remaining clauses will remain in full force and effect.
+
+## 14.2 If we fail to insist that you perform any of your obligations under these Terms, or if we do not enforce our rights against you, or if we delay in doing so, that will not mean that we have waived our rights against you and will not mean that you do not have to comply with those obligations. If we do waive a default by you, we will only do so in writing, and that will not mean that we will automatically waive any later default by you.
+
+## 14.3 The rights, powers and remedies provided in these Terms are (except as expressly provided) cumulative and not exclusive of any rights, powers and remedies provided by applicable laws or otherwise.
+
+## 14.4 These Terms constitute the entire agreement between you and us in relation to their subject matter, and supersede and extinguish all prior agreements, arrangements, promises, undertakings, proposals, warranties, representations, and understandings between made by you and us (“Pre-Contractual Statements”), whether oral or written, in relation to such subject matter. You and us both acknowledge that in entering into these Terms we do not rely on any Pre-Contractual Statement made by or on behalf of the other party (whether made innocently or negligently) in relation to the subject matter of these Terms, other than those which are set out expressly in these Terms, and you and us hereby waive all rights and remedies which might otherwise be available to you or us in relation to such Pre-Contractual Statements. Nothing in this clause 14.4 shall limit or exclude the liability of either party arising out of its pre-contract fraudulent misrepresentation or fraudulent concealment.
+
+## 14.5 You may not assign, subcontract or otherwise transfer or dispose of any of your rights or obligations under these Terms without our prior written consent. We may freely assign, subcontract or otherwise transfer or dispose of any of our rights or obligations under these Terms without your consent.
+
+## 14.6 Nothing in these Terms shall create or be deemed to create a partnership between you and us.
+
+## 14.7 Save as expressly set out in these Terms, a person who is not a party to these Terms has no right to enforce any part of these Terms.
