+++ mocks/data/terms/business/15_law.md
@@ -1,0 +1,5 @@
+# 15. GOVERNING LAW AND JURISDICTION
+
+## 15.1 These Terms and any dispute or claim arising out of or in connection with it or its subject matter or formation (including non-contractual disputes or claims) shall be governed by and construed in accordance with the law of England and Wales.
+
+## 15.2 You irrevocably agree that the courts of England and Wales have exclusive jurisdiction to settle any dispute or claim that arises out of or in connection with this Agreement or its subject matter or formation (including non-contractual disputes or claims).
