+++ mocks/data/terms/business/16_contact.md
@@ -1,0 +1,9 @@
+# 16. CONTACTING US
+
+## 16.1 SOURSD Helpdesk is available from [Insert] to [Insert] on [business days].
+
+## 16.2 Should you have any reasons for a complaint, we will endeavour to resolve the issue and avoid any re-occurrence in the future.
+
+## 16.3 You can contact us by clicking [here] or by using the following details:
+
+### Email address: [enquiries@soursd.org](enquiries@soursd.org)
