+++ mocks/data/terms/business/1_understanding.md
@@ -1,0 +1,27 @@
+# 1. UNDERSTANDING THESE TERMS OF USE
+
+## 1.1 These terms of use (the “Terms”) describe how you may access and use SOURSD. By accepting these Terms as part of the Account (defined below) registration process, you acknowledge and agree to these Terms. If you do not agree to any of these Terms you must not use SOURSD.
+
+## 1.2 These Terms cover access to and use of SOURSD by users (“Users”) acting:
+
+### 1.2.1 on behalf of themselves in order to access sensitive data on TREs or SDEs for research purposes (a “Researcher”);
+
+### 1.2.2 as a delegate on behalf of a university, research institute, commercial organisation or administrative body in order to access sensitive data on TREs or SDEs for research purposes (in which case such organisation will be a “Research Organisation”); and
+
+### 1.2.3 acting as a delegate on behalf of organisations that manage a TRE or SDE in order to set rules to validate whether a Researcher or Research Organisation meets the criteria to access the TRE or SDE that you manage (in which case such organisation will be a “Data Custodian”).
+
+## 1.3 In these Terms, when we refer to "you" or "your," we mean:
+
+### 1.3.1 you, the person accessing or using SOURSD as a User; and
+
+### 1.3.2 where applicable, the Research Organisation or Data Custodian on whose behalf you are accessing SOURSD.
+
+## 1.4 If you are acting on behalf of a Research Organisation or Data Custodian when you access and use SOURSD, you represent and warrant that:
+
+### 1.4.1 you have full legal authority to bind such Research Organisation or Data Custodian (as applicable); and
+
+### 1.4.2 you agree to these Terms on behalf of the Research Organisation or Data Custodian (as applicable) that you represent.
+
+## 1.5 We will only use your personal data through SOURSD in accordance with our privacy policy (available [insert URL]).
+
+## 1.6 Where you are acting on behalf of a Research Organisation or Data Custodian you agree that HDR UK and the Research Organisation and/or Data Custodian are independent controllers for the processing of personal data associated with SOURSD and that HDR UK and the applicable Research Organisation/Data Custodian will comply with applicable data protection legislation such as the UK GDPR.
