+++ mocks/data/terms/business/2_usage_researcher.md
@@ -1,0 +1,19 @@
+# 2. USE OF SOURSD BY RESEARCHERS
+
+## 2.1 If you are a Researcher, then this clause 2 will apply to you (and clauses 3 and 4 will not apply to you).
+
+## 2.2 Subject to these Terms, you may access and use SOURSD for the following purposes:
+
+### 2.2.1 to populate data fields in your SOURSD profile such as data in relation to your identity, your affiliations, your experience, your training and your location;
+
+### 2.2.2 to view the data that you have provided in your Account;
+
+### 2.2.3 to request that Research Organisations add you as an affiliate;
+
+### 2.2.4 to view any data access decisions made by Data Custodians in your Account based on the data that you have provided; and
+
+### 2.2.5 to view projects you have been associated with by a Data Custodian,
+
+(collectively, the “Researcher Permitted Purpose”).
+
+## 2.3 We grant you a non-exclusive, non-transferable, right (without the right to grant sub-licences) to use and access SOURSD solely for the Researcher Permitted Purpose.
