+++ mocks/data/terms/business/3_usage_org.md
@@ -1,0 +1,25 @@
+# 3. USE OF SOURSD BY RESEARCH ORGANISATIONS
+
+## 3.1 If you are a Research Organisation (or a User acting on behalf of a Research Organisation), then this clause 3 will apply to you (and clauses 2 and 4 will not apply to you).
+
+## 3.2 Subject to these Terms, you may access and use SOURSD for the following purposes:
+
+### 3.2.1 to populate data fields in your SOURSD profile such as data in relation to your organisation’s name and address, your organisation identifiers, your data security compliance, details of your nominated Senior Responsible Officer (“SRO”) and any of your authorised delegates;
+
+### 3.2.2 to view the data that you have provided in your Account;
+
+### 3.2.3 to invite Researchers to set up Accounts;
+
+### 3.2.4 to view any data access decisions made by Data Custodians in your Account based on the data that you have provided;
+
+### 3.2.5 to view, copy, download and export a list of your affiliated Users; and
+
+### 3.2.6 to view Researchers and projects associated with your Research Organisation by a Data Custodian,
+
+(collectively, the “Research Organisation Permitted Purpose”).
+
+## 3.3 We grant you a non-exclusive, non-transferable, right (without the right to grant sub-licences) to use and access SOURSD solely for the Research Organisation Permitted Purpose.
+
+## 3.4 As a Research Organisation, you shall identify an SRO who will be responsible for your Account and nominating various authorised delegates. Each SRO shall follow the SRO Declaration process as set out in SOURSD in order for their Account to be verified and made active.
+
+## 3.5 Where you are based outside of the UK or EEA in a country that has not been recognised by the UK government as providing an adequate level of protection for personal data then you consent to HDR UK transferring your personal data to the relevant Research Organisation for them to verify your association with them and for the Research Organisation to review its affiliated Users and projects.
