+++ mocks/data/terms/business/4_usage_custodian.md
@@ -1,0 +1,41 @@
+# 4. USE OF SOURSD BY DATA CUSTODIANS
+
+## 4.1 If you are a Data Custodian (or a User acting on behalf of a Data Custodian), then this clause 4 will apply to you (and clauses 2 and 3 will not apply to you).
+
+## 4.2 Subject to these Terms, you may access and use SOURSD for the following purposes:
+
+### 4.2.1 to populate data fields in your SOURSD profile such as data in relation to your organisation, your authorised administrators and your ‘approvers’. An ‘approver’ is an individual who records any decisions made regarding the Researcher’s ‘safety’ (as defined in the Five Safes Framework) and access to specific projects;
+
+### 4.2.2 to configure rules in relation to the criteria for accessing your TRE or SDE (as applicable);
+
+### 4.2.3 to configure validation checks to assess properties of Researcher and Research Organisation profiles against TRE or SDE user validation policies;
+
+### 4.2.4 to populate details about specific projects;
+
+### 4.2.5 to view, analyse, copy, download and export data inputted into SOURSD by Researchers and Research Organisations;
+
+### 4.2.6 to assess the ‘safety’ of Researchers and Research Organisations (as defined in the Five Safes Framework);
+
+### 4.2.7 to notify Researchers and Research Organisations when they have been validated or added to a project;
+
+### 4.2.8 to populate your Data Use Register if you have an account to the Health Data Research Gateway, subject to you having opted to utilise this functionality; and
+
+### 4.2.9 to allow Users to sign into your TRE or SRE with their SOURSD Account log-in details or digital identifiers,
+
+(collectively, the “Data Custodian Permitted Purpose”).
+
+## 4.3 We grant you a non-exclusive, non-transferable, right (without the right to grant sub-licences) to use and access SOURSD solely for the Data Custodian Permitted Purpose.
+
+## 4.4 As a Data Custodian, you shall identify a main contact who is responsible for your Account and for nominating administrators and approvers. An ‘approver’ is an individual who records any decisions made regarding the Researcher’s ‘safety’ (as defined in the Five Safes Framework) and access to specific projects.
+
+## 4.5 You represent and warrant that:
+
+### 4.5.1 you are accredited under one of the following:
+
+#### 4.5.1.1 the Digital Economy Act (DEA) 2017; or
+
+#### 4.5.1.2 the Scottish Safe Haven Charter;
+
+### 4.5.2 you are an SDE within the NSH SDE Network; or
+
+### 4.5.3 your TRE or SDE meets all the mandatory controls listed in the Standard Architecture for Trusted Research Environments (SATRE) available at: https://satre-specification.readthedocs.io/en/stable/.
