+++ mocks/data/terms/business/5_account.md
@@ -1,0 +1,23 @@
+# 5. YOUR ACCOUNT AND PASSWORD
+
+## 5.1 You will need to register an account with us on SOURSD in order to access the services available on SOURSD (“Account”).
+
+## 5.2 In order to register an Account, you must be aged 18 or over. You will be asked to provide certain information (such as your name and e-mail address) and create a password. You must treat the password (and any unique identification code provided as part of account registration) as confidential and you must not disclose it to any third party.
+
+## 5.3 You agree that:
+
+### 5.3.1 you must not allow anyone else to use your Account password or Account login details;
+
+### 5.3.2 all the information that you provide to us in connection with your Account is complete and accurate;
+
+### 5.3.3 you are the person whose details you have provided;
+
+### 5.3.4 you will take all reasonable steps to keep your Account secure and to prevent any fraudulent use of it; and
+
+### 5.3.5 you will notify us immediately if there are any changes to the information you have provided to us.
+
+## 5.4 We have the right to disable any Accounts and/or passwords at any time if, in our reasonable opinion, you have failed to comply with any of the provisions of these Terms.
+
+## 5.5 If you know or suspect that anyone other than you knows your Account login details and/or password, or you become aware of any transactions on your Account which have not been validly authorised, you must immediately notify us at dataprotection@hdruk.ac.uk.
+
+## 5.6 You are responsible for any unauthorised use of your Account.
