+++ mocks/data/terms/business/6_acceptable_use.md
@@ -1,0 +1,31 @@
+# 6. ACCEPTABLE USE
+
+## 6.1 You agree not to:
+
+### 6.1.1 use SOURSD in any way that breaches these Terms or any applicable law or regulation;
+
+### 6.1.2 use SOURSD for the purpose of making hiring decisions;
+
+### 6.1.3 copy, modify or otherwise reproduce or re-sell any part of SOURSD; or
+
+### 6.1.4 do any act or thing that might damage, disrupt or otherwise interfere with the operation of SOURSD or any equipment, network or software used in operating SOURSD.
+
+## 6.2 You must ensure that the data you supply to SOURSD must comply with the following rules:
+
+### 6.2.1 it must not be obscene, abusive, offensive or racist and it must not promote or propose hatred or physical harm against anyone;
+
+### 6.2.2 it must not harass or bully another person;
+
+### 6.2.3 it must be complete, accurate, true and honest (so far as you know);
+
+### 6.2.4 it must not be defamatory of anyone;
+
+### 6.2.5 it must not infringe the rights or privacy of anyone else;
+
+### 6.2.6 it must not contain someone else’s personal details or confidential information relating to other people; and
+
+### 6.2.7 it must not promote or condone terrorism, violence or illegal behaviour.
+
+## 6.3 We reserve the right to refuse to accept or cease to use any Input Data (as defined below) supplied by you that we consider contravenes these Terms.
+
+## 6.4 You must not misuse SOURSD by knowingly introducing viruses, trojans, worms, logic bombs or other material which is malicious or technologically harmful. You must not attempt to gain unauthorised access to SOURSD, the server on which SOURSD is stored or any server, computer or database connected to SOURSD.
