+++ mocks/data/terms/business/7_intellectual_property.md
@@ -1,0 +1,5 @@
+# 7. INTELLECTUAL PROPERTY
+
+## 7.1 In these terms “Intellectual Property Rights” means: (a) patents, inventions, designs, copyright and related rights, database rights, trade-marks, service marks and trade names (whether registered or unregistered), and rights to apply for registration; (b) proprietary rights in domain names; (c) knowhow and confidential information; (d) applications, extensions and renewals in relation to any of these rights; and (e) all other rights of a similar nature or having an equivalent effect which currently exist anywhere in the world.
+
+## 7.2 Save for Input Data (for which clause 8 below shall apply), we are the owner or licensee of all Intellectual Property Rights in SOURSD and its content. Except as expressly permitted herein, these Terms do not grant you any rights to, or in, any Intellectual Property Rights in respect of SOURSD.
