+++ mocks/data/terms/business/8_input_data.md
@@ -1,0 +1,17 @@
+# 8. INPUT DATA
+
+## 8.1 In these Terms, “Input Data” means any data that you input into SOURSD.
+
+## 8.2 You will retain all rights, title and interest in and to all of your Input Data, subject to any pre-existing rights that third parties may have in its content. You shall at all times have sole responsibility for the legality, reliability, integrity, accuracy and quality of the Input Data and for ensuring that its use does not infringe the rights of any third parties. If your Input Data contains material and/or content owned or generated by a third party, you should ensure that you have all the necessary licences, permissions and/or consents from that third party to use such material and/or before inputting such Input Data into SOURSD.
+
+## 8.3 You grant us a non-exclusive, irrevocable, royalty-free, sub-licensable, fully paid up, perpetual and worldwide licence to access, use, copy, distribute and prepare derivative works of your Input Data for the purposes of providing SOURSD.
+
+## 8.4 You warrant that use of the Input Data in accordance with these Terms will not:
+
+### 8.4.1 breach any laws, statutes or regulations;
+
+### 8.4.2 infringe the Intellectual Property Rights or other legal rights of any person; or
+
+### 8.4.3 give rise to any cause of action against us, in each case in any jurisdiction and under any applicable law.
+
+## 8.5 In the event of any loss or damage to Input Data, we will use reasonable commercial endeavours to restore the lost or damaged Input Data from the latest back-up of such Input Data (where such back-ups are made by us). This will be your sole and exclusive remedy. We shall not be responsible for any loss, destruction, alteration or disclosure of Input Data caused by you or any third party (except those third parties sub-contracted by us to perform services, in which case our liability shall be subject to the limitations and exclusions set out in clause 9).
