+++ mocks/data/terms/business/9_liability.md
@@ -1,0 +1,15 @@
+# 9. OUR LIABILITY
+
+## 9.1 Nothing in these Terms excludes or limits our liability for:
+
+### 9.1.1 death or personal injury caused by our negligence;
+
+### 9.1.2 fraud or fraudulent misrepresentation; or
+
+### 9.1.3 any matter in respect of which it would be unlawful for us to exclude or restrict our liability.
+
+## 9.2 Subject to clause 9.1, our total liability to you arising out of or in connection with these Terms and/or the provision of SOURSD, whether in contract (including under any indemnity), misrepresentation (whether tortious or statutory), tort (including negligence), breach of statutory duty, or otherwise shall be limited to £500.
+
+## 9.3 Subject to clause 9.1, in no event shall we be liable to you for any loss of profits, loss of revenue, loss of contracts, damage to goodwill, failure to realise anticipated savings or for any indirect or consequential loss, whether arising from negligence, breach of contract or otherwise, even if we were aware of the possibility that such loss or damage might be incurred by you.
+
+## 9.4 You shall indemnify and hold us harmless against any losses, costs, liabilities and expenses suffered or incurred by us and/or our affiliates as a result of any breach of these Terms (including any breach of any warranty in these terms).
