+++ package.json
@@ -141,6 +141,7 @@
     "miragejs": "^0.1.48",
     "next-router-mock": "^0.9.13",
     "prettier": "^3.2.5",
+    "raw-loader": "^4.0.2",
     "react-syntax-highlighter": "^15.5.0",
     "semantic-release": "^24.2.1",
     "storybook": "^8.0.5",
